+++ packages/core/legos/src/sources.ts
@@ -8,6 +8,7 @@ import * as amberdata from '@chainlink/amberdata-adapter'
 import * as anyblock from '@chainlink/anyblock-adapter'
 import * as ap_election from '@chainlink/ap-election-adapter'
 import * as armanino from '@chainlink/armanino-adapter'
+import * as avalanche_platform from '@chainlink/avalanche-platform-adapter'
 import * as bank_frick from '@chainlink/bank-frick-adapter'
 import * as bea from '@chainlink/bea-adapter'
 import * as binance from '@chainlink/binance-adapter'
@@ -89,6 +90,7 @@ import * as ipfs from '@chainlink/ipfs-adapter'
 import * as jpegd from '@chainlink/jpegd-adapter'
 import * as json_adapter from '@chainlink/json-rpc-adapter'
 import * as kaiko from '@chainlink/kaiko-adapter'
+import * as kaiko_test from '@chainlink/kaiko-test-adapter'
 import * as layer2_sequencer_health from '@chainlink/layer2-sequencer-health-adapter'
 import * as lcx from '@chainlink/lcx-adapter'
 import * as lido from '@chainlink/lido-adapter'
@@ -101,7 +103,7 @@ import * as metalsapi from '@chainlink/metalsapi-adapter'
 import * as mock_ea from '@chainlink/mock-ea-adapter'
 import * as mycryptoapi from '@chainlink/mycryptoapi-adapter'
 import * as ncfx from '@chainlink/ncfx-adapter'
-import * as ncfx_test from '@chainlink/ncfx-test-adapter'
+import * as nft_blue_chip from '@chainlink/nft-blue-chip-adapter'
 import * as nikkei from '@chainlink/nikkei-adapter'
 import * as nomics from '@chainlink/nomics-adapter'
 import * as oilpriceapi from '@chainlink/oilpriceapi-adapter'
@@ -156,6 +158,7 @@ export default {
   anyblock,
   ap_election,
   armanino,
+  avalanche_platform,
   bank_frick,
   bea,
   binance,
@@ -237,6 +240,7 @@ export default {
   jpegd,
   json_adapter,
   kaiko,
+  kaiko_test,
   layer2_sequencer_health,
   lcx,
   lido,
@@ -249,7 +253,7 @@ export default {
   mock_ea,
   mycryptoapi,
   ncfx,
-  ncfx_test,
+  nft_blue_chip,
   nikkei,
   nomics,
   oilpriceapi,