+++ packages/core/reference-data-reader/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @chainlink/ea-reference-data-reader
 
+## 1.3.26
+
+### Patch Changes
+
+- Updated dependencies [838c9d927]
+  - @chainlink/ea-bootstrap@2.25.2
+
 ## 1.3.25
 
 ### Patch Changes