+++ packages/sources/1forge/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @chainlink/1forge-adapter
 
+## 1.6.28
+
+### Patch Changes
+
+- Updated dependencies [838c9d927]
+  - @chainlink/ea-bootstrap@2.25.2
+  - @chainlink/ea-test-helpers@1.4.1
+
 ## 1.6.27
 
 ### Patch Changes