+++ packages/sources/amberdata/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @chainlink/amberdata-adapter
 
+## 1.8.20
+
+### Patch Changes
+
+- Updated dependencies [838c9d927]
+  - @chainlink/ea-bootstrap@2.25.2
+  - @chainlink/ea-factories@1.2.26
+  - @chainlink/ea-test-helpers@1.4.1
+
 ## 1.8.19
 
 ### Patch Changes