+++ packages/sources/chain-reserve-wallet/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @chainlink/chain-reserve-wallet-adapter
 
+## 3.0.11
+
+### Patch Changes
+
+- Updated dependencies [838c9d927]
+  - @chainlink/ea-bootstrap@2.25.2
+  - @chainlink/ea-test-helpers@1.4.1
+
 ## 3.0.10
 
 ### Patch Changes